pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Agent workspace development baseline for tech_blog_monitor modernization."
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.14.0",
    "apscheduler>=3.10.4",
    "fastapi>=0.115.0",
    "feedparser>=6.0.11",
    "loguru>=0.7.2",
    "opentelemetry-exporter-otlp-proto-http>=1.31.0",
    "opentelemetry-sdk>=1.31.0",
    "pgvector>=0.3.6",
    "playwright>=1.53.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "psycopg[binary]>=3.2.0",
    "PyYAML>=6.0.2",
    "requests>=2.32.0",
    "SQLAlchemy>=2.0.36",
    "trafilatura>=1.12.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.11.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["common", "products", "runtime"]
